refactors seektest and adds more formats
